Plugin class gets injected with plugin info
This replaces the initial module key parameter in the ctor.
Additionally, the plugin now merges the initial container with the contents of the services file,
as well as the injected plugin info with the contents of the config file.
These services and config entries are included in the setup container.