Remove patch number from min EDD version
Having it there used to cause `version_compare()` to give incorrect result. Seems to be a bug, which manifests because `EDD_VERSION` also is missing the patch number. It does not appear practical to look for the bug report to link to, because this is likely due to a very old PHP version being used.